+++ src/main/java/org/accounting/system/repositories/project/ProjectRepository.java
@@ -20,6 +20,7 @@ import org.accounting.system.entities.MetricDefinition;
 import org.accounting.system.entities.acl.RoleAccessControl;
 import org.accounting.system.entities.authorization.Role;
 import org.accounting.system.entities.projections.InstallationProjection;
+import org.accounting.system.entities.projections.MetricReportProjection;
 import org.accounting.system.entities.projections.MongoQuery;
 import org.accounting.system.entities.projections.ProjectReport;
 import org.accounting.system.entities.projections.ProviderReport;
@@ -44,6 +45,7 @@ import org.bson.Document;
 import org.bson.conversions.Bson;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
@@ -599,7 +601,6 @@ public class ProjectRepository extends ProjectModulator {
 
         var group = Aggregates.group(groupId, Accumulators.sum("totalValue", "$value"));
 
-
         var extractFields = Aggregates.addFields(
                 new Field<>("providerId", "$_id.providerId"),
                 new Field<>("installationId", "$_id.installationId"),
@@ -669,7 +670,8 @@ public class ProjectRepository extends ProjectModulator {
                         .append("infrastructure", "$infrastructure")
                         .append("resource", "$resource")
                         .append("data", "$data")
-                )
+                ),
+                Accumulators.push("allMetrics", "$data")
         );
 
         var lookupProvider = Aggregates.lookup("Provider", "_id", "_id", "provider");
@@ -682,18 +684,79 @@ public class ProjectRepository extends ProjectModulator {
                 Projections.computed("logo", new Document("$ifNull", List.of("$provider.logo", ""))),
                 Projections.computed("name", "$provider.name"),
                 Projections.computed("website", new Document("$ifNull", List.of("$provider.website", ""))),
-                Projections.include("data")
+                Projections.include("data"),
+                Projections.computed("aggregatedMetrics", new Document("$reduce", new Document()
+                        .append("input", "$allMetrics")
+                        .append("initialValue", new ArrayList<>())
+                        .append("in", new Document("$concatArrays", Arrays.asList("$$value", "$$this")))
+                ))));
+
+        var agg = Aggregates.unwind("$aggregatedMetrics");
+
+        var gagg = Aggregates.group(new Document("provider_id", "$provider_id").append("metric_definition_id", "$aggregatedMetrics.metricDefinitionId"), Accumulators.sum("totalValue", "$aggregatedMetrics.totalValue"),
+                Accumulators.first("metricDefinitionId", "$aggregatedMetrics.metricDefinitionId"),
+                Accumulators.first("metricName", "$aggregatedMetrics.metricName"),
+                Accumulators.first("metricDescription", "$aggregatedMetrics.metricDescription"),
+                Accumulators.first("metricType", "$aggregatedMetrics.metricType"),
+                Accumulators.first("unitType", "$aggregatedMetrics.unitType"),
+                Accumulators.first("name", "$name"),
+                Accumulators.first("website", "$website"),
+                Accumulators.first("abbreviation", "$abbreviation"),
+                Accumulators.first("logo", "$logo"),
+                Accumulators.first("data", "$data")
+        );
+
+        var gpafggg = Aggregates.project(Projections.fields(
+                Projections.computed("provider_id", "$_id.provider_id"),
+                Projections.include("name", "website", "abbreviation", "logo", "data", "unitType", "metricType", "metricDescription", "metricName", "metricDefinitionId", "totalValue")
         ));
 
+        var gpagg = Aggregates.group(
+                "$provider_id",
+                Accumulators.first("name", "$name"),
+                Accumulators.first("website", "$website"),
+                Accumulators.first("abbreviation", "$abbreviation"),
+                Accumulators.first("logo", "$logo"),
+                Accumulators.first("data", "$data"),
+                Accumulators.first("provider_id", "$provider_id"),
+                Accumulators.push("aggregatedMetrics", new Document()
+                        .append("metricDefinitionId", "$_id.metric_definition_id")
+                        .append("metricName", "$metricName")
+                        .append("metricDescription", "$metricDescription")
+                        .append("unitType", "$unitType")
+                        .append("metricType", "$metricType")
+                        .append("totalValue", "$totalValue")
+                )
+        );
 
         var data = metricRepository.getMongoCollection()
                 .aggregate(List.of(regex, addField, group, extractFields,
                         lookup, unwind, projection, finalGroup,
                         lookupInstallation, unwindInstallation,
                         finalProjection, groupByProvider, lookupProvider, unwindProvider,
-                        finalProviderProjection), ProviderReport.class).into(new ArrayList<>());
+                        finalProviderProjection, agg, gagg, gpafggg, gpagg), ProviderReport.class).into(new ArrayList<>());
 
         var project = fetchById(projectId);
+
+        var aggregatedMetrics = data
+                .stream()
+                .flatMap(provider->provider.aggregatedMetrics.stream())
+                .collect(Collectors.toMap(k-> k.metricDefinitionId,
+                m -> {
+                    var copy = new MetricReportProjection();
+                    copy.metricDefinitionId = m.metricDefinitionId;
+                    copy.metricName = m.metricName;
+                    copy.metricDescription = m.metricDescription;
+                    copy.unitType = m.unitType;
+                    copy.metricType = m.metricType;
+                    copy.totalValue = m.totalValue;
+                    return copy;
+                },
+                (m1, m2) -> {
+                    m1.totalValue = m1.totalValue + m2.totalValue;
+                    return m1;
+                }
+        ));
 
         var report = new ProjectReport();
         report.data = data;
@@ -703,6 +766,7 @@ public class ProjectRepository extends ProjectModulator {
         report.title = project.title;
         report.startDate = project.startDate;
         report.endDate = project.endDate;
+        report.aggregatedMetrics = new ArrayList<>(aggregatedMetrics.values());
 
         return  report;
     }