+++ CHANGELOG.md
@@ -31,6 +31,11 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the
 - [#233](https://github.com/ARGOeu/argo-accounting/pull/233) - ACC-370 Remove
   data concerning permissions in reports.
 
+## Unreleased - Added
+
+- [#234](https://github.com/ARGOeu/argo-accounting/pull/234) - ACC-373 Add
+  "group by" feature in reports.
+
 ## 1.7.0 - 2025-07-25
 
 ---